@philpearl 2020-03-15
Yep plan to sort this out soon. Was waiting for the preceding PR to merge

On Sun, 15 Mar 2020, 00:43 Alfred Sorten Wolf, <notifications@github.com>
wrote:

> It has been 14 days with no activity and the awaiting response label was
> assigned. Is this PR still valid? Assigning the stalled label. Please
> comment to reassure me that this is still being worked on.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/tensorflow/tensorflow/pull/36898#issuecomment-599152720>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AADPPXDZV52DTTLN33IDBOLRHQQCRANCNFSM4KXZUAKA>
> .
>