@philpearl 2020-03-16
Hopefully the subsequent PR to reinstate it (
https://github.com/tensorflow/tensorflow/pull/37190) hasn't been rolled
back.

On Mon, 16 Mar 2020 at 18:37, Rajeshwar Reddy T <notifications@github.com>
wrote:

> @philpearl <https://github.com/philpearl> #36578
> <https://github.com/tensorflow/tensorflow/pull/36578> got rolled back due
> to internal failures , how do we proceed with this PR ?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/tensorflow/tensorflow/pull/36898#issuecomment-599698189>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AADPPXANVGGPHGIE3DPLT2DRHZWYFANCNFSM4KXZUAKA>
> .
>