@philpearl 2020-03-16
Dammit the re-instating PR (https://github.com/tensorflow/tensorflow/pull/37190) was rolled back. It just says "due to internal failure" with no further explanation.

At this point I think I give up. I believe in these changes, but dealing with this codebase with its missing generated files and mysterious extra background testing processes is just too painfull